requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.2.1
jinja2==3.1.4
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.4.0
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.8
requests==2.32.3
httpx==0.27.0
aiohttp==3.10.5
python-dotenv==1.0.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
